minor refactoring of bt-interface.sh script

+++ recipes-multimedia/seamless-point/files/bt-interface.sh
@@ -4,16 +4,15 @@ BT_MAC_DEVICES="bluetoothctl devices | grep "Device" | cut -d ' ' -f 2"
 BT_DEFAULT_SYMBOL="🔈"
 BT_SERVICE_NAME="bt"
 
-SHAIRPORT_SYNC="shairport-sync"
-BASEDIR="/tmp"
+TEMPDIR="/tmp"
 
-generate_random_port() {
-    local mac_address="$1"
+generate_arbitrary_port() {
+    local mac="$1"
     local min_port=1     # Minimum port number
-    local max_port=100   # Maximum port number
+    local max_port=255   # Maximum port number
     local range=$((max_port - min_port + 1))
 
-    _hash="$(echo -n "$mac_address" | md5sum | cut -d' ' -f1)"
+    _hash="$(echo -n "$mac" | md5sum | cut -d' ' -f1)"
     hash_decimal="$((16#${_hash}))"
     port_number="$((min_port + (hash_decimal % range)))"
 
@@ -23,12 +22,8 @@ generate_random_port() {
 process_mac() {
     local mac="${1}"
 
-    # Convert MAC address to lowercase and remove colons
-    local processed_mac
-    processed_mac=$(echo "${mac}" | tr '[:upper:]' '[:lower:]' | tr -d ':')
-
-    # Output the processed MAC address
-    echo "${processed_mac}"
+    # Convert MAC address to lowercase plus remove colons
+    echo "${mac}" | tr '[:upper:]' '[:lower:]' | tr -d ':'
 }
 
 stop_airplay() {
@@ -37,17 +32,16 @@ stop_airplay() {
 
     # Drop running shairport-sync services to start them after connection
     for version in v2 v1; do
-        systemctl is-active --quiet ${SHAIRPORT_SYNC}-${version}@${alsadev}.service
+        systemctl is-active --quiet shairport-sync-${version}@${alsadev}.service
         if [ ${?} -eq 0 ]; then
-            systemctl stop ${SHAIRPORT_SYNC}-${version}@${alsadev}.service
-            # Drop env files
-            [  -r ${BASEDIR}/${version}-${alsadev}-extraopts.env ] && \
-                rm -f ${BASEDIR}/${version}-${alsadev}-extraopts.env
+            systemctl stop shairport-sync-${version}@${alsadev}.service
 
-            for file in ${BASEDIR}/${SHAIRPORT_SYNC}-${version}-*-bt.conf; do
-                if [ -e "${file}" ]; then
-                    unlink "${file}"
-                fi
+            # Drop environment files
+            [  -r ${TEMPDIR}/${version}-${alsadev}-extraopts.env ] && \
+                rm -f ${TEMPDIR}/${version}-${alsadev}-extraopts.env
+
+            for file in ${TEMPDIR}/shairport-sync-${version}-*-bt.conf; do
+                [ -e "${file}" ] && unlink "${file}"
             done
         fi
     done
@@ -62,7 +56,7 @@ stop_services() {
     done
 }
 
-update_config() {
+update_sp_config() {
     local config_file_path="${1}"
     local new_device_id_offset="${2}"
 
@@ -76,10 +70,10 @@ process_connected() {
 
     # Fetch device name from Bluetooth attributes
     for version in v1 v2; do
-        offset="$(generate_random_port ${alsadev})"
+        offset="$(generate_arbitrary_port ${alsadev})"
 
         if [ ${version} == "v1" ]; then
-            port="--port $((5000+ ${offset}))"
+            port="--port $((5000 + ${offset}))"
         elif [ ${version} == "v2" ]; then
             port="--port $((7000 + ${offset}))"
         else
@@ -89,27 +83,28 @@ process_connected() {
         # Process routine only in case shairport-sync service is not running
         systemctl is-active --quiet ${SHAIRPORT_SYNC}-${version}@${alsadev}.service
         if [ ${?} -ne 0 ]; then
-            ln -fsr /etc/${SHAIRPORT_SYNC}-${version}-bt.conf ${BASEDIR}/${SHAIRPORT_SYNC}-${version}-${alsadev}-bt.conf
-            update_config ${BASEDIR}/${SHAIRPORT_SYNC}-${version}-${alsadev}-bt.conf ${offset}
+            ln -fsr /etc/shairport-sync-${version}-bt.conf ${TEMPDIR}/shairport-sync-${version}-${alsadev}-bt.conf
+            update_sp_config ${TEMPDIR}/shairport-sync-${version}-${alsadev}-bt.conf ${offset}
 
             # Update the device name from Bluetooth attributes, otherwise keep default "Bluetooth Speaker" name
             local device_name="$(bluetoothctl info ${mac} | grep Name: | awk -F': ' '{print $2}')"
             if [ -n "${device_name}" ]; then
-                # =JBL Speaker
-                # =f85c7de6c3aa
-                # =7002
-                # =v1
-                # =-c /tmp/shairport-sync shairport-sync-bt.conf 7002 --name JBL Speaker --output alsa -d f85c7de6c3aa
-                local env=${BASEDIR}/${version}-${alsadev}-extraopts.env
+                local env=${TEMPDIR}/${version}-${alsadev}-extraopts.env
                 local devname="${BT_DEFAULT_SYMBOL} ${device_name}"
+
+                # JBL Speaker
                 echo "DEVICE_NAME=\"${devname}\"" > ${env}
+                # Sound card name: e.g. f85c7de6c3aa
                 echo "ALSA_DEVICE=\"${alsadev}\"" >> ${env}
+                # 7xxx 5xxx
                 echo "PORT=\"${port}\"" >> ${env}
+                # v1 v2
                 echo "VERSION=\"${version}\"" >> ${env}
-                echo "OPTS=\"-c ${BASEDIR}/${SHAIRPORT_SYNC}-${version}-${alsadev}-bt.conf ${port} --name '${devname}' --output=alsa -- -d ${alsadev}\"" >> ${env}
+                # -c /tmp/shairport-sync shairport-sync-bt.conf 7002 --name JBL Speaker --output alsa -d f85c7de6c3aa
+                echo "OPTS=\"-c ${TEMPDIR}/shairport-sync-${version}-${alsadev}-bt.conf ${port} --name '${devname}' --output=alsa -- -d ${alsadev}\"" >> ${env}
             fi
 
-            systemctl start ${SHAIRPORT_SYNC}-${version}@${alsadev}.service
+            systemctl start shairport-sync-${version}@${alsadev}.service
         fi
     done
 }
